+++ src/main/scala/com/twitter/chill/RichKryo.scala
@@ -20,7 +20,7 @@ import com.esotericsoftware.kryo.Kryo
 import com.esotericsoftware.kryo.{ Serializer => KSerializer }
 import com.esotericsoftware.kryo.io.{ Input, Output }
 
-import com.twitter.bijection.{ Bijection, Bufferable }
+import com.twitter.bijection.{ Bufferable, ImplicitBijection, Injection }
 
 import scala.collection.mutable.Builder
 
@@ -32,40 +32,36 @@ class RichKryo(k: Kryo) {
   def alreadyRegistered[T](implicit cmf: ClassManifest[T]): Boolean =
     k.getClassResolver.getRegistration(cmf.erasure) != null
 
-  def bijectionForClass[T](implicit bij: Bijection[T, Array[Byte]], cmf: ClassManifest[T]):
-    Kryo = {
-      k.register(cmf.erasure, BijectiveSerializer.asKryo[T])
-      k
-    }
+  def injectionForClass[T](implicit inj: Injection[T, Array[Byte]], cmf: ClassManifest[T]): Kryo = {
+    k.register(cmf.erasure, InjectiveSerializer.asKryo[T])
+    k
+  }
 
-  def bijectionForClasses(pairs: TraversableOnce[BijectionPair[_]]): Kryo = {
-    pairs.foreach { pair: BijectionPair[_] =>
+  def injectionForClasses(pairs: TraversableOnce[InjectionPair[_]]): Kryo = {
+    pairs.foreach { pair: InjectionPair[_] =>
       if (!alreadyRegistered(ClassManifest.fromClass(pair.klass))) {
-        val serializer = BijectiveSerializer.asKryo(pair.bijection)
+        val serializer = InjectiveSerializer.asKryo(pair.injection)
         k.register(pair.klass, serializer)
       } else {
-        System.err.printf("%s is already registered in registerBijections.",
-                          Array[String](pair.klass.getName))
+        System.err.printf("%s is already registered in injectionForClasses.", pair.klass.getName)
       }
     }
     k
   }
 
-  def bijectionForSubclass[T](implicit bij: Bijection[T, Array[Byte]], cmf: ClassManifest[T]):
-    Kryo = {
-      k.addDefaultSerializer(cmf.erasure, BijectiveSerializer.asKryo[T])
-      k
-    }
+  def injectionForSubclass[T](implicit inj: Injection[T, Array[Byte]], cmf: ClassManifest[T]): Kryo = {
+    k.addDefaultSerializer(cmf.erasure, InjectiveSerializer.asKryo[T])
+    k
+  }
 
-  def bijectionForSubclasses(pairs: TraversableOnce[BijectionPair[_]]): Kryo = {
-    pairs.foreach { pair: BijectionPair[_] =>
+  def injectionForSubclasses(pairs: TraversableOnce[InjectionPair[_]]): Kryo = {
+    pairs.foreach { pair: InjectionPair[_] =>
       if (!alreadyRegistered(ClassManifest.fromClass(pair.klass))) {
-        val serializer = BijectiveSerializer.asKryo(pair.bijection)
+        val serializer = InjectiveSerializer.asKryo(pair.injection)
         k.addDefaultSerializer(pair.klass, serializer)
         k.register(pair.klass)
       } else {
-        System.err.printf("%s is already registered in registerBijectionDefaults.",
-                          Array[String](pair.klass.getName))
+        System.err.printf("%s is already registered in injectionForSubclasses.", pair.klass.getName)
       }
     }
     k
@@ -98,8 +94,7 @@ class RichKryo(k: Kryo) {
 
   /** B has to already be registered, then use the KSerializer[B] to create KSerialzer[A]
    */
-  def forClassViaBijection[A,B]
-    (implicit bij: Bijection[A,B], acmf: ClassManifest[A], bcmf: ClassManifest[B]): Kryo = {
+  def forClassViaBijection[A,B](implicit bij: ImplicitBijection[A,B], acmf: ClassManifest[A], bcmf: ClassManifest[B]): Kryo = {
     val kserb = k.getSerializer(bcmf.erasure).asInstanceOf[KSerializer[B]]
     k.register(acmf.erasure, KryoSerializer.viaBijection[A,B](kserb))
     k